autoIndent.sh Linux bugbix

+++ useful/autoIndent.sh
@@ -49,19 +49,21 @@ for elem in ${@:2}; do
     printf "$(loadColor 1)Indent all $elem...$(loadColor 0)\n";
     filesList=$(find $1 -iname \*$elem);
     for file in $filesList; do
-        tmpFile=$(getRandomName);
-        (cp $file $tmpFile;
-            echo -n "" > $file;
-            while IFS='' read -r line || [[ -n "$line" ]]; do
-		echo "$line" >> $file;
-            done < $tmpFile
-            rm $tmpFile;
-            emacs -nw -q --batch $file --eval "(mark-whole-buffer)" --eval "(indent-region (point-min) (point-max) nil)" --eval "(save-buffer)" --kill > /dev/null 2> /dev/null; 
-            if [ -f $file'~' ]; then
-                rm $file'~';
-            fi
-            printf "\t$file $(loadColor 1 32)Ok$(loadColor 0)\n"; ) &
-        taskList+=($!);
+	if [ ! -L $file ]; then
+            tmpFile=$(getRandomName);
+            (cp $file $tmpFile;
+		echo -n "" > $file;
+		while IFS='' read -r line || [[ -n "$line" ]]; do
+		    echo "$line" >> $file;
+		done < $tmpFile
+		rm $tmpFile;
+		emacs -nw -q --batch $file --eval "(mark-whole-buffer)" --eval "(indent-region (point-min) (point-max) nil)" --eval "(save-buffer)" --kill > /dev/null 2> /dev/null; 
+		if [ -f $file'~' ]; then
+                    rm $file'~';
+		fi
+		printf "\t$file $(loadColor 1 32)Ok$(loadColor 0)\n"; ) &
+            taskList+=($!);
+	fi
     done 
     for task in ${taskList[@]}; do
         wait $task;
